@@ -1,83 +1,226 @@
 {
-  "name": "mapboxgl-react-redux",
-  "private": true,
-  "version": "0.0.1",
-  "description": "mapboxgl-react-redux - Generated by generator-react-webpack",
-  "main": "src/index.js",
+  "name": "react-slingshot",
+  "version": "7.0.0",
+  "description": "Starter kit for creating apps with React and Redux",
+  "engines": {
+    "npm": ">=3"
+  },
   "scripts": {
-    "clean": "rimraf dist/*",
-    "copy": "copyfiles -f ./src/index.html ./dist && copyfiles -u 1 \"./src/static/**\" ./dist",
-    "dist": "npm run clean && npm run copy && webpack --progress --bail --env dist -p",
-    "lint": "eslint ./src",
-    "posttest": "npm run lint",
-    "release:major": "npm version prerelease && git push --follow-tags && npm publish --tag beta",
-    "release:minor": "npm version prerelease && git push --follow-tags && npm publish --tag beta",
-    "release:patch": "npm version prerelease && git push --follow-tags && npm publish --tag beta",
-    "serve:dev": "webpack-dev-server --open --env dev",
-    "serve:dist": "webpack-dev-server --open --env dist -p --progress",
-    "start": "npm run serve:dev",
-    "test": "cross-env NODE_ENV=test karma start",
-    "test:watch": "cross-env NODE_ENV=test karma start --autoWatch=true --singleRun=false --reporters=mocha,coverage"
+    "preinstall": "node tools/nodeVersionCheck.js",
+    "setup": "node tools/setup/setupMessage.js && npm install && node tools/setup/setup.js",
+    "remove-demo": "babel-node tools/removeDemo.js",
+    "start-message": "babel-node tools/startMessage.js",
+    "prestart": "npm-run-all --parallel start-message",
+    "start": "npm-run-all --parallel test:watch open:src lint:watch",
+    "open:src": "babel-node tools/srcServer.js",
+    "open:dist": "babel-node tools/distServer.js",
+    "lint": "esw webpack.config.* src tools --color",
+    "lint:watch": "npm run lint -- --watch",
+    "clean-dist": "npm run remove-dist && mkdir dist",
+    "remove-dist": "rimraf ./dist",
+    "prebuild": "npm run clean-dist && npm run lint && npm run test",
+    "build": "babel-node tools/build.js && npm run open:dist",
+    "test": "jest",
+    "test:CI": "babel-node tools/testCi.js",
+    "test:cover": "npm run test -- --coverage ",
+    "test:cover:CI": "npm run test:CI -- --coverage && cat ./coverage/lcov.info | node_modules/coveralls/bin/coveralls.js",
+    "test:watch": "npm run test -- --watch",
+    "open:cover": "npm run test:cover && opn ./coverage/lcov-report/index.html",
+    "analyze-bundle": "babel-node ./tools/analyzeBundle.js"
   },
-  "repository": "",
-  "keywords": [],
-  "author": "Your name here",
-  "devDependencies": {
-    "babel-core": "^6.7.6",
-    "babel-eslint": "^7.1.0",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-istanbul": "^3.0.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-rest-spread": "^6.8.0",
-    "babel-polyfill": "^6.9.0",
-    "babel-preset-airbnb": "^2.0.0",
-    "babel-preset-es2015-native-modules": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "chai": "^3.5.0",
-    "copyfiles": "^1.0.0",
-    "css-loader": "^0.26.0",
-    "enzyme": "^2.2.0",
-    "eslint": "^3.0.0",
-    "eslint-config-airbnb": "^13.0.0",
-    "eslint-loader": "^1.3.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-jsx-a11y": "^2.2.0",
-    "eslint-plugin-react": "^6.0.0",
-    "file-loader": "^0.9.0",
-    "karma": "^1.0.0",
-    "karma-chai": "^0.1.0",
-    "karma-coverage": "^1.0.0",
-    "karma-junit-reporter": "^1.0.0",
-    "karma-mocha": "^1.0.1",
-    "karma-mocha-reporter": "^2.0.3",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "karma-sinon": "^1.0.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.7.0",
-    "mapbox-gl": "^0.38.0",
-    "mocha": "^3.0.0",
-    "node-sass": "^3.7.0",
-    "null-loader": "^0.1.1",
-    "phantomjs-prebuilt": "^2.1.7",
-    "react-addons-test-utils": "^15.0.1",
-    "rimraf": "^2.5.2",
-    "sass-loader": "^3.2.0",
-    "sinon": "^1.17.3",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack-dev-server": "^2.1.0-beta"
-  },
+  "author": "Cory House",
+  "license": "MIT",
   "dependencies": {
-    "cross-env": "^3.1.0",
-    "mapbox-gl": "^0.38.0",
-    "react": "^15.0.1",
-    "react-css-modules": "^3.7.10",
-    "react-dom": "^15.0.1",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-map-gl": "^2.0.3",
-    "react-mapbox-gl": "^2.0.6",
-    "react-redux": "^5.0.5",
-    "redux": "^3.7.1",
-    "webpack": "^2.1.0-beta.6"
+    "object-assign": "4.1.0",
+    "react": "15.5.4",
+    "react-dom": "15.5.4",
+    "react-redux": "5.0.5",
+    "react-router": "3.0.5",
+    "react-router-redux": "4.0.8",
+    "redux": "3.6.0",
+    "redux-thunk": "2.1.0"
+  },
+  "devDependencies": {
+    "autoprefixer": "7.1.2",
+    "babel-cli": "6.24.1",
+    "babel-core": "6.25.0",
+    "babel-eslint": "7.2.3",
+    "babel-jest": "20.0.3",
+    "babel-loader": "7.1.1",
+    "babel-plugin-transform-react-constant-elements": "6.23.0",
+    "babel-plugin-transform-react-remove-prop-types": "0.4.6",
+    "babel-polyfill": "6.23.0",
+    "babel-preset-env": "1.6.0",
+    "babel-preset-react": "6.24.1",
+    "babel-preset-react-hmre": "1.1.1",
+    "babel-preset-stage-1": "6.24.1",
+    "browser-sync": "2.18.12",
+    "chalk": "2.0.1",
+    "connect-history-api-fallback": "1.3.0",
+    "coveralls": "2.13.1",
+    "css-loader": "0.28.4",
+    "enzyme": "2.9.1",
+    "eslint": "4.2.0",
+    "eslint-plugin-import": "2.7.0",
+    "eslint-plugin-react": "7.1.0",
+    "eslint-watch": "3.1.2",
+    "extract-text-webpack-plugin": "3.0.0-rc.2",
+    "file-loader": "0.11.2",
+    "html-webpack-plugin": "2.29.0",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "20.0.4",
+    "json-loader": "0.5.4",
+    "mockdate": "2.0.1",
+    "node-sass": "4.5.3",
+    "npm-run-all": "4.0.2",
+    "opn-cli": "3.1.0",
+    "postcss-loader": "1.2.1",
+    "prompt": "1.0.0",
+    "prop-types": "15.5.10",
+    "react-hot-loader": "3.0.0-beta.6",
+    "react-test-renderer": "15.5.4",
+    "redux-immutable-state-invariant": "2.0.0",
+    "replace": "0.3.0",
+    "rimraf": "2.6.1",
+    "sass-loader": "6.0.6",
+    "style-loader": "0.18.2",
+    "url-loader": "0.5.9",
+    "webpack": "3.1.0",
+    "webpack-bundle-analyzer": "2.8.2",
+    "webpack-dev-middleware": "1.11.0",
+    "webpack-hot-middleware": "2.18.2",
+    "webpack-md5-hash": "0.0.5"
+  },
+  "keywords": [
+    "react",
+    "reactjs",
+    "react-router",
+    "hot",
+    "reload",
+    "hmr",
+    "live",
+    "edit",
+    "webpack",
+    "redux",
+    "flux",
+    "boilerplate",
+    "starter"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/coryhouse/react-slingshot"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(css|scss)$": "identity-obj-proxy",
+      "^.+\\.(gif|ttf|eot|svg|woff|woff2|ico)$": "<rootDir>/tools/fileMock.js"
+    }
+  },
+  "babel": {
+    "presets": ["react", "stage-1"],
+    "env": {
+      "development": {
+        "presets": ["env", "react-hmre"]
+      },
+      "production": {
+        "presets": [
+          [
+            "env",
+            {
+              "es2015": {
+                "modules": false
+              }
+            }
+          ]
+        ],
+        "plugins": [
+          "transform-react-constant-elements",
+          "transform-react-remove-prop-types"
+        ]
+      },
+      "test": {
+        "presets": ["env"]
+      }
+    }
+  },
+  "eslintConfig": {
+    "root": true,
+    "extends": [
+      "eslint:recommended",
+      "plugin:import/errors",
+      "plugin:import/warnings"
+    ],
+    "plugins": ["react"],
+    "parser": "babel-eslint",
+    "parserOptions": {
+      "ecmaVersion": 6,
+      "sourceType": "module",
+      "ecmaFeatures": {
+        "jsx": true,
+        "experimentalObjectRestSpread": true
+      }
+    },
+    "env": {
+      "es6": true,
+      "browser": true,
+      "node": true,
+      "jquery": true,
+      "jest": true
+    },
+    "rules": {
+      "quotes": 0,
+      "no-console": 1,
+      "no-debugger": 1,
+      "no-var": 1,
+      "semi": [1, "always"],
+      "no-trailing-spaces": 0,
+      "eol-last": 0,
+      "no-underscore-dangle": 0,
+      "no-alert": 0,
+      "no-lone-blocks": 0,
+      "jsx-quotes": 1,
+      "react/display-name": [
+        1,
+        {
+          "ignoreTranspilerName": false
+        }
+      ],
+      "react/forbid-prop-types": [
+        1,
+        {
+          "forbid": ["any"]
+        }
+      ],
+      "react/jsx-boolean-value": 0,
+      "react/jsx-closing-bracket-location": 0,
+      "react/jsx-curly-spacing": 1,
+      "react/jsx-indent-props": 0,
+      "react/jsx-key": 1,
+      "react/jsx-max-props-per-line": 0,
+      "react/jsx-no-bind": 0,
+      "react/jsx-no-duplicate-props": 1,
+      "react/jsx-no-literals": 0,
+      "react/jsx-no-undef": 1,
+      "react/jsx-pascal-case": 1,
+      "react/jsx-sort-prop-types": 0,
+      "react/jsx-sort-props": 0,
+      "react/jsx-uses-react": 1,
+      "react/jsx-uses-vars": 1,
+      "react/jsx-wrap-multilines": 1,
+      "react/no-danger": 1,
+      "react/no-did-mount-set-state": 1,
+      "react/no-did-update-set-state": 1,
+      "react/no-direct-mutation-state": 1,
+      "react/no-multi-comp": 1,
+      "react/no-set-state": 0,
+      "react/no-unknown-property": 1,
+      "react/prefer-es6-class": 1,
+      "react/prop-types": 1,
+      "react/react-in-jsx-scope": 1,
+      "import/extensions": 1,
+      "react/self-closing-comp": 1,
+      "react/sort-comp": 1
+    },
+    "globals": {}
   }
 }
