@@ -1,19 +1,28 @@
 import React from 'react';
-import MapContainer from './mapContainer/MapContainer';
-import './app.css';
+import PropTypes from 'prop-types';
+import { Link, IndexLink } from 'react-router';
 
-class AppComponent extends React.Component {
-
+// This is a class-based component because the current
+// version of hot reloading won't hot reload a stateless
+// component at the top-level.
+class App extends React.Component {
   render() {
     return (
-      <div className="top">
-        <MapContainer />
+      <div>
+        <IndexLink to="/">Home</IndexLink>
+        {' | '}
+        <Link to="/fuel-savings">Demo App</Link>
+        {' | '}
+        <Link to="/about">About</Link>
+        <br/>
+        {this.props.children}
       </div>
     );
   }
 }
 
-AppComponent.defaultProps = {
+App.propTypes = {
+  children: PropTypes.element
 };
 
-export default AppComponent;
+export default App;
